requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
aiohttp==3.8.4
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.3
sqlalchemy==2.0.15
python-dateutil==2.8.2
pytz==2023.3